SingleStockPage uses themes
